libpdbg/host: Actually check the return value of open() in host_pib_probe()
host_pib_probe() will allocate memory for its private bookkeeping and
then store a file descriptor that it open()s. The check that open()
succeeded looks at the pointer where the file descriptor was put, not
the actual value of it.

Looks like pointers never evaluate to being less than zero - which is
handy but has also meant this bug won't be found easily.

While at it, fix the leak of the private storage if open() does fail.