libpdbg/htm: Enable NCU and set single dispatch/instruction mode
It isn't clear if chtm needs to be enabled before or after setting
single dispatch single instruction mode in the core. It also isn't
clear if the NCU enable should be done before of after core enable.

This fixes problems related with chtm being stuck in "INIT" state and
failing to become "READY".